Added covid command and little refactor

+++ WeatherBot.Domain/Models/Weather/WeatherModel.cs
@@ -4,7 +4,7 @@
     {
         public float temp { get; set; }
         public float feels_like { get; set; }
-        public int temp_min { get; set; }
+        public float temp_min { get; set; }
         public float temp_max { get; set; }
         public int pressure { get; set; }
         public int humidity { get; set; }
